requirements.txt
numpy==1.20.3
pytorch ==1.8.1
streamlit==0.1.8.8
pillow==8.2.0 
pandas==1.2.4    
matplotlib==3.4.2        